Fix the show more lines toogle
When the file name contained dot the id of the preview section
and the data attribute used in show more toggle link were not identical.
One used file name without dots, the other with dots.

I also changed the replace method to replace all dots, not only the
first one.